requirements.txt
botbuilder-integration-aiohttp>=4.14.0
aiohttp  # Asegúrate de verificar la versión correcta
python-dotenv
botbuilder-core
botbuilder-schema
convertapi
